rename: 팀 대시보드용 하이라이트 파일명 수정
